requirements.txt
python-dotenv>=1.0.1
pandas>=2.2.2
sentence-transformers>=3.0.1
faiss-cpu>=1.8.0.post1
numpy>=1.26.4
requests>=2.32.3
google-generativeai>=0.8.2
fastapi>=0.115.0
uvicorn[standard]>=0.30.6
pydantic>=2.9.2

